% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_le.R
\name{clean_le}
\alias{clean_le}
\title{Read and clean latent heat flux data in a FLUXNET2015 .csv file.}
\usage{
clean_le(f)
}
\arguments{
\item{f}{Path to .csv file.}
}
\value{
Data frame containing timeseries of latent heat flux values with
low quality data removed.
}
\description{
Read and clean latent heat flux data in a FLUXNET2015 .csv file.
}
\examples{
\dontrun{
f <- system.file('extdata',
  "FLX_AR-SLu_FLUXNET2015_FULLSET_HH_2009-2011_1-4.csv",
  package='fluxcleanr'
)
clean_le(f)
}
}
