% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_le.R
\name{process_le}
\alias{process_le}
\title{Process all raw FLUXNET2015 data in a directory.}
\usage{
process_le(flux_dir, out_dir, avg_daily = TRUE)
}
\arguments{
\item{flux_dir}{Directory containing raw fluxnet data.}

\item{out_dir}{Directory to save processed fluxnet data to.}

\item{avg_daily}{boolean - Return the average daily latent heat flux for each
site (TRUE) or keep the data at half-hourly and hourly intervals (FALSE).}
}
\value{
Data frame where daily missing values are gap filled and
}
\description{
Process all raw FLUXNET2015 data in a directory.
}
\examples{
\dontrun{
process_le('path/to/flux/dir', '/path/to/save/dir', avg_daily = TRUE)
}
}
