% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_le.R, R/filter_le.R
\name{clean_le}
\alias{clean_le}
\title{Read and clean latent heat flux data in a FLUXNET2015 .csv file.}
\usage{
clean_le(f)

clean_le(f)
}
\arguments{
\item{f}{Path to .csv file.}
}
\value{
Data frame containing timeseries of latent heat flux values.

Data frame containing timeseries of latent heat flux values.
}
\description{
Read and clean latent heat flux data in a FLUXNET2015 .csv file.

Filter out missing latent heat flux such that all data from all flux towers have at least 2 years of complete data.
}
\examples{
clean_le('http://files.ntsg.umt.edu/data/colin_data/fluxcleanr/hh_ex.csv')
clean_le('http://files.ntsg.umt.edu/data/colin_data/fluxcleanr/hh_ex.csv')
}
